Formatting change

+++ Projects/SupplyChainProject/ItemManager.sol
@@ -6,11 +6,8 @@ import "./Item.sol";
 import "./Ownable.sol";
 
 contract ItemManager is Ownable {
-    enum SupplyChainSteps {
-        Created,
-        Paid,
-        Delivered
-    }
+
+    enum SupplyChainSteps{Created, Paid, Delivered}
 
     struct S_Item {
         Item item;
@@ -18,19 +15,12 @@ contract ItemManager is Ownable {
         SupplyChainSteps step;
     }
 
-    mapping(uint256 => S_Item) public items;
-    uint256 index;
+    mapping(uint => S_Item) public items;
+    uint index;
 
-    event SupplyChainEvent(
-        uint256 index,
-        SupplyChainSteps step,
-        address itemAddress
-    );
+    event SupplyChainEvent(uint index, SupplyChainSteps step, address itemAddress);
 
-    function createItem(string memory _identifier, uint256 _price)
-        public
-        onlyOwner
-    {
+    function createItem(string memory _identifier, uint _price) public onlyOwner {
         Item item = new Item(this, index, _price);
         items[index].item = item;
         items[index].identifier = _identifier;
@@ -38,45 +28,28 @@ contract ItemManager is Ownable {
 
         emit SupplyChainEvent(index, items[index].step, address(item));
 
-        index++;
+        index++; 
     }
 
     // Acess instance variables of external contracts as function
     // Not as variable
     // Because solidity automatically creates getter funciton for public variables
-    function triggerPayment(uint256 _index) public payable {
+    function triggerPayment(uint _index) public payable {
         Item item = items[_index].item;
 
-        require(
-            address(item) == msg.sender,
-            "Only Item is allowed to update itself"
-        );
+        require(address(item) == msg.sender, "Only Item is allowed to update itself");
         require(item.price() == msg.value, "Full payments only");
-        require(
-            items[_index].step == SupplyChainSteps.Created,
-            "Item further in supply chain"
-        );
+        require(items[_index].step == SupplyChainSteps.Created, "Item further in supply chain");
 
         items[_index].step = SupplyChainSteps.Paid;
-
-        emit SupplyChainEvent(
-            _index,
-            items[_index].step,
-            address(items[_index].item)
-        );
+        
+        emit SupplyChainEvent(_index, items[_index].step, address(items[_index].item));
     }
 
-    function triggerDelivery(uint256 _index) public onlyOwner {
-        require(
-            items[_index].step == SupplyChainSteps.Paid,
-            "Item is further in Supply Chain"
-        );
+    function triggerDelivery(uint _index) public onlyOwner {
+        require(items[_index].step == SupplyChainSteps.Paid, "Item is further in Supply Chain");
         items[_index].step = SupplyChainSteps.Delivered;
 
-        emit SupplyChainEvent(
-            _index,
-            items[_index].step,
-            address(items[_index].item)
-        );
+        emit SupplyChainEvent(_index, items[_index].step, address(items[_index].item));
     }
-}
+ }
